# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://127.0.0.1:8000'
paths:
...
  /api/purchase:
    post:
      summary: ''
      operationId: postApiPurchase
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                product_id:
                  type: string
                  description: 'The <code>id</code> of an existing record in the products table.'
                  example: id
                  nullable: false
                quantity:
                  type: integer
                  description: 'Must be at least 1.'
                  example: 65
                  nullable: false
              required:
                - product_id
      security: []
  '/api/purchase/{id}/status':
    patch:
      summary: ''
      operationId: patchApiPurchaseIdStatus
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  description: ''
                  example: completed
                  nullable: false
                  enum:
                    - completed
                    - cancelled
              required:
                - status
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the purchase.'
        example: et
        required: true
        schema:
          type: string
...
  /api/stores:
    get:
      summary: ''
      operationId: getApiStores
      description: ''
      parameters: []
      responses:
        500:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Server Error'
                properties:
                  message:
                    type: string
                    example: 'Server Error'
      tags:
        - Endpoints
      security: []
    post:
      summary: ''
      operationId: postApiStores
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: abtqczpbo
                  nullable: false
                type:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: akehrjfpfgklayrhrof
                  nullable: false
                country:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: okacudxrg
                  nullable: false
                address:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: dbkyyunsmywqoseu
                  nullable: false
                phone:
                  type: string
                  description: 'Must not be greater than 20 characters.'
                  example: jaufxpblbdgixilwux
                  nullable: false
                image:
                  type: string
                  format: binary
                  description: 'Must be an image. Must not be greater than 2048 kilobytes.'
                  nullable: false
              required:
                - name
                - type
                - country
                - address
                - phone
      security: []
...
  '/api/stores/{store_id}':
    post:
      summary: ''
      operationId: postApiStoresStore_id
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: false
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: slklgqonxniuxldrfpx
                  nullable: true
                type:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: tpgobrbyghwipdabmu
                  nullable: true
                country:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: tfiffhwzhqhcxmwrielz
                  nullable: true
                address:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: dyjovthlyqdzyivkgr
                  nullable: true
                phone:
                  type: string
                  description: 'Must not be greater than 20 characters.'
                  example: tdok
                  nullable: true
                image:
                  type: string
                  format: binary
                  description: 'Must be an image. Must not be greater than 2048 kilobytes.'
                  nullable: true
      security: []
    get:
      summary: ''
      operationId: getApiStoresStore_id
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  success: true
                  message: ok
                  data:
                    id: 1
                    name: matjaaaaaaaar
                    type: consequatur
                    country: Anguilla
                    address: '6767 Upton Orchard Suite 737Port Orpha, ND 18441-2630'
                    phone: 802.573.2698
                    image: 'https://cdn.britannica.com/16/204716-050-8BB76BE8/Walmart-store-Mountain-View-California.jpg'
                    points: 134
                    products: []
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: ok
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: matjaaaaaaaar
                      type:
                        type: string
                        example: consequatur
                      country:
                        type: string
                        example: Anguilla
                      address:
                        type: string
                        example: '6767 Upton Orchard Suite 737Port Orpha, ND 18441-2630'
                      phone:
                        type: string
                        example: 802.573.2698
                      image:
                        type: string
                        example: 'https://cdn.britannica.com/16/204716-050-8BB76BE8/Walmart-store-Mountain-View-California.jpg'
                      points:
                        type: integer
                        example: 134
                      products:
                        type: array
                        example: []
      tags:
        - Endpoints
      security: []
    delete:
      summary: ''
      operationId: deleteApiStoresStore_id
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: store_id
        description: 'The ID of the store.'
        example: 1
        required: true
        schema:
          type: integer
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []